backend/requirements.txt
# FastAPI 核心
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
aiosqlite==0.19.0
SQLAlchemy[asyncio]==2.0.23

# 文件处理
python-multipart==0.0.6
Pillow==10.1.0

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 环境变量和配置
python-dotenv==1.0.0
pydantic-settings==2.1.0

# 开发工具 (可选)
pytest==7.4.3
pytest-asyncio==0.21.1
jinja2==3.1.2
